refactor: Invitation 도메인을 클린 아키텍쳐 패키지로 변경 #37
