+++ src/main/java/site/praytogether/pray_together/domain/invitation/application/InvitationApplicationService.java
@@ -5,12 +5,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import site.praytogether.pray_together.domain.base.MessageResponse;
-import site.praytogether.pray_together.domain.invitation.dto.InvitationCreateRequest;
-import site.praytogether.pray_together.domain.invitation.dto.InvitationInfoScrollResponse;
-import site.praytogether.pray_together.domain.invitation.dto.InvitationStatusUpdateRequest;
-import site.praytogether.pray_together.domain.invitation.model.Invitation;
-import site.praytogether.pray_together.domain.invitation.model.InvitationInfo;
-import site.praytogether.pray_together.domain.invitation.service.InvitationService;
+import site.praytogether.pray_together.domain.invitation.presentation.dto.InvitationCreateRequest;
+import site.praytogether.pray_together.domain.invitation.presentation.dto.InvitationInfoScrollResponse;
+import site.praytogether.pray_together.domain.invitation.presentation.dto.InvitationStatusUpdateRequest;
+import site.praytogether.pray_together.domain.invitation.domain.Invitation;
+import site.praytogether.pray_together.domain.invitation.domain.InvitationInfo;
+import site.praytogether.pray_together.domain.invitation.domain.service.InvitationService;
 import site.praytogether.pray_together.domain.member.model.Member;
 import site.praytogether.pray_together.domain.member.service.MemberService;
 import site.praytogether.pray_together.domain.member_room.service.MemberRoomService;
