+++ src/test/java/site/praytogether/pray_together/test_config/IntegrateTest.java
@@ -4,14 +4,13 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
 import site.praytogether.pray_together.domain.friend.domain.friend_invitation.FriendInvitationRepository;
 import site.praytogether.pray_together.domain.friend.domain.friendship.FriendshipRepository;
-import site.praytogether.pray_together.domain.invitation.repository.InvitationRepository;
+import site.praytogether.pray_together.domain.invitation.domain.repository.InvitationRepository;
 import site.praytogether.pray_together.domain.member.repository.MemberRepository;
 import site.praytogether.pray_together.domain.member_room.repository.MemberRoomRepository;
 import site.praytogether.pray_together.domain.notification.repository.PrayerCompletionNotificationRepository;
